Fold RlePlus into BitField
