+++ utils/bitfield/src/writer.rs
